+++ SerialPrograms/Source/PokemonLA/Programs/Farming/PokemonLA_LeapGrinder.cpp
@@ -177,7 +177,7 @@ bool LeapGrinder::run_iteration(SingleSwitchProgramEnvironment& env, BotBaseCont
     int ret = run_until(
         env.console, context,
         [&](BotBaseContext& context){
-            route(env, env.console, context, POKEMON);
+            route(env, env.console, context, (size_t)POKEMON);
         },
         {{shiny_detector}}
     );
@@ -255,7 +255,7 @@ bool LeapGrinder::run_iteration(SingleSwitchProgramEnvironment& env, BotBaseCont
 
     env.console.log("Remaining Leaps:" + std::to_string(LEAPS - stats.leaps));
 
-    return_to_jubilife(env, env.console, context, POKEMON);
+    return_to_jubilife(env, env.console, context, (size_t)POKEMON);
 
     if (stats.leaps == LEAPS){
         return true;
